Second[Agent] upgrade next.js
