+++ app/list/add-or-remove-items/page.js
@@ -1,8 +1,9 @@
+'use client'
 
 
 import { useState, useEffect } from 'react';
 import Router, { useRouter } from 'next/router';
-import Head from 'next/head';
+import { Metadata } from 'next/head';
 
 import Header from 'parts/Header';
 import NotFound from 'parts/NotFound';
@@ -204,9 +205,9 @@ const AddOrRemoveItems = ({
   if (status === STATUSES.RESOLVED) {
     return (
       <>
-        <Head>
+        <Metadata>
           <title>{movies.name}</title>
-        </Head>
+        </Metadata>
         <PageWrapper>
           <PaddingWrapper>
             <Header
