+++ app/list/page.js
@@ -1,7 +1,6 @@
-
+'use client'
 
 import { useEffect, useState } from 'react';
-import Head from 'next/head';
 import { useRouter } from 'next/router';
 import { animateScroll as scroll } from 'react-scroll';
 
@@ -9,61 +8,63 @@ import Header from 'parts/Header';
 import NotFound from 'parts/NotFound';
 import PageWrapper from 'parts/PageWrapper';
 import PaddingWrapper from 'parts/PaddingWrapper';
-import MyTMDBLists from 'components/MyTMDBLists';
+import ListActions from 'containers/ListActions';
+import MovieList from 'components/MovieList';
 import Loader from 'components/UI/Loader';
-import withAuth from 'utils/hocs/withAuth';
+import { useAuth } from 'utils/hocs/AuthProvider';
 import { TMDB_API_NEW_VERSION, TMDB_IMAGE_BASE_URL } from 'config/tmdb';
 import QUERY_PARAMS from 'utils/constants/query-params';
 import STATUSES from 'utils/constants/statuses';
 import tmdbAPI from 'services/tmdbAPI';
 
-const MyLists = ({
-  accountId,
-  accessToken
-}) => {
+const List = () => {
   const { query } = useRouter();
+
+  const { accessToken } = useAuth();
 
   const [status, setStatus] = useState(STATUSES.IDLE);
   // TODO: could handle errors
   const [error, setError] = useState(null);
 
-  const [myLists, setMyLists] = useState(null);
+  const [movies, setMovies] = useState(null);
 
   const page = Number(query[QUERY_PARAMS.PAGE]);
+  const listId = query[QUERY_PARAMS.ID];
 
   useEffect(() => {
     (async () => {
       if (!page) return;
-      if (!accountId) return;
-      if (!accessToken) return;
+      if (!listId) return;
 
       scroll.scrollToTop({smooth: true});
 
       try {
         setStatus(STATUSES.PENDING);
-        const response = await tmdbAPI.get(`/${TMDB_API_NEW_VERSION}/account/${accountId}/lists`, {
-          headers: {
-            'Authorization': `Bearer ${accessToken}`
-          },
+        const headers = accessToken ? {
+          'Authorization': `Bearer ${accessToken}`
+        } : {};
+
+        const response = await tmdbAPI.get(`/${TMDB_API_NEW_VERSION}/list/${listId}`, {
+          headers,
           params: {
             page
           }
         });
-        const myLists = response.data;
-        setMyLists(myLists);
+        const movies = response.data;
+        setMovies(movies);
       } catch (error) {
-        console.log('[MyLists useEffect] error => ', error);
+        console.log('[List useEffect] error => ', error);
         setStatus(STATUSES.REJECTED);
         setError(error);
       }
     })();
-  }, [page, accountId, accessToken]);
+  }, [page, listId, accessToken]);
 
   useEffect(() => {
-    if (!myLists) return;
+    if (!movies) return;
 
     setStatus(STATUSES.RESOLVED);
-  }, [myLists]);
+  }, [movies]);
 
   if (status === STATUSES.IDLE || status === STATUSES.PENDING) {
     return <Loader />;
@@ -74,24 +75,33 @@ const MyLists = ({
     return (
       <NotFound
         title='Sorry!'
-        subtitle={error?.message ?? 'There were no my lists...'} />
+        subtitle={error?.message ?? 'We can\'t find the page you\'re looking for...'} />
     );
   }
 
   if (status === STATUSES.RESOLVED) {
     return (
       <>
-        <Head>
-          <title>My Lists</title>
-        </Head>
+        <Metadata title={movies.name ?? 'List'} />
         <PageWrapper>
           <PaddingWrapper>
             <Header
-              title='My Lists'
-              subtitle='TMDB' />
-            <MyTMDBLists
-              myLists={myLists}
-              baseUrl={TMDB_IMAGE_BASE_URL} />
+              title={movies.name ?? 'No name'}
+              subtitle={movies.description ?? 'No description'} />
+            <ListActions
+              listId={listId}
+              page={page}
+              listName={movies.name}
+              creatorAccountId={movies?.created_by?.id} />
+            {movies.total_results === 0 ? (
+              <NotFound
+                title='Sorry!'
+                subtitle='There were no items...' />
+            ) : (
+              <MovieList
+                movies={movies}
+                baseUrl={TMDB_IMAGE_BASE_URL} />
+            )}
           </PaddingWrapper>
         </PageWrapper>
       </>
@@ -99,4 +109,4 @@ const MyLists = ({
   }
 };
 
-export default withAuth(MyLists);
+export default List;
