+++ components/PosterLink/index.js
@@ -1,4 +1,5 @@
 
+import React from "react";
 import Link from 'next/link';
 
 import withTheme from 'utils/hocs/withTheme';
@@ -11,10 +12,7 @@ const PosterLink = React.forwardRef(({
   ...rest
 }, ref) => (
   <>
-    <Link
-      passHref
-      as={as}
-      href={href}>
+    <Link passHref as={as} href={href} legacyBehavior>
       <a
         ref={ref}
         {...rest}>
