+++ components/UI/TheSelectSearch/index.js
@@ -1,5 +1,7 @@
 
 
+import React from "react";
+
 import SelectSearch from 'react-select-search/dist/cjs';
 import clsx from 'clsx';
