+++ containers/DarkModeToggle/index.js
@@ -1,7 +1,4 @@
-
-
-// MEMO: inspired by https://web.dev/prefers-color-scheme/#the-lessdark-mode-togglegreater-custom-element
-import Head from 'next/head';
+import { Metadata } from 'components/Metadata';
 import clsx from 'clsx';
 import useDarkMode from 'use-dark-mode';
 
@@ -21,83 +18,16 @@ const DarkModeToggle = ({
 
   return (
     <>
-      <Head>
-        {darkMode.value ? (
-          <>
-            {/* TODO: block for now for toggle experience of the favicon depending on the light/dark mode */}
-            {/* <link
-              rel='icon'
-              href='/dark-favicon.ico' /> */}
-            <meta name="viewport" content="width=device-width, initial-scale=1" />
-            <link
-              rel='apple-touch-icon'
-              sizes='180x180'
-              href='/dark-apple-touch-icon.png' />
-            <link
-              rel='icon'
-              type='image/png'
-              sizes='32x32'
-              href='/dark-favicon-32x32.png' />
-            <link
-              rel='icon'
-              type='image/png'
-              sizes='16x16'
-              href='/dark-favicon-16x16.png' />
-            <link
-              rel='manifest'
-              href='/dark-manifest.webmanifest' />
-            <link
-              rel='mask-icon'
-              href='/dark-safari-pinned-tab.svg'
-              color='#5bbad5' />
-            <meta
-              name='msapplication-TileColor'
-              content='#da532c' />
-            {/* TODO: hardcoded */}
-            <meta name='theme-color' content='#fafafa' />
-            <meta
-              name='msapplication-config'
-              content='/dark-browserconfig.xml' />
-          </>
-        ) : (
-          <>
-            {/* TODO: block for now for toggle experience of the favicon depending on the light/dark mode */}
-            {/* <link
-              rel='icon'
-              href='/light-favicon.ico' /> */}
-            <meta name="viewport" content="width=device-width, initial-scale=1" />
-            <link
-              rel='apple-touch-icon'
-              sizes='180x180'
-              href='/light-apple-touch-icon.png' />
-            <link
-              rel='icon'
-              type='image/png'
-              sizes='32x32'
-              href='/light-favicon-32x32.png' />
-            <link
-              rel='icon'
-              type='image/png'
-              sizes='16x16'
-              href='/light-favicon-16x16.png' />
-            <link
-              rel='manifest'
-              href='/light-manifest.webmanifest' />
-            <link
-              rel='mask-icon'
-              href='/light-safari-pinned-tab.svg'
-              color='#5bbad5' />
-            <meta
-              name='msapplication-TileColor'
-              content='#da532c' />
-            {/* TODO: hardcoded */}
-            <meta name='theme-color' content='#303030' />
-            <meta
-              name='msapplication-config'
-              content='/light-browserconfig.xml' />
-          </>
-        )}
-      </Head>
+      <Metadata
+        title={darkMode.value ? 'Dark Mode' : 'Light Mode'}
+        description={darkMode.value ? 'Switch to light mode' : 'Switch to dark mode'}
+        favicon={darkMode.value ? '/dark-favicon.ico' : '/light-favicon.ico'}
+        appleTouchIcon={darkMode.value ? '/dark-apple-touch-icon.png' : '/light-apple-touch-icon.png'}
+        manifest={darkMode.value ? '/dark-manifest.webmanifest' : '/light-manifest.webmanifest'}
+        maskIcon={darkMode.value ? '/dark-safari-pinned-tab.svg' : '/light-safari-pinned-tab.svg'}
+        msapplicationConfig={darkMode.value ? '/dark-browserconfig.xml' : '/light-browserconfig.xml'}
+        themeColor={darkMode.value ? '#fafafa' : '#303030'}
+      />
       <div className={clsx('dark-mode-toggle', className)}>
         <button
           type='button'
