+++ containers/SearchBar/Form/index.js
@@ -1,5 +1,7 @@
 
 
+import React from "react";
+
 const Form = React.forwardRef(({
   opened,
   theme,
