+++ containers/SearchBar/Input/index.js
@@ -1,4 +1,5 @@
 
+import React from "react";
 const Input = React.forwardRef(({
   opened,
   theme,
