+++ package.json
@@ -15,30 +15,30 @@
     "analyze:browser": "cross-env BUNDLE_ANALYZE=browser next build"
   },
   "dependencies": {
-    "@artsy/fresnel": "^1.2.0",
-    "@loadable/component": "^5.13.1",
-    "clsx": "^1.1.1",
-    "next": "^12.2.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-glider": "^2.0.1",
-    "react-lazyload": "^3.0.0",
-    "react-modal-video": "^1.2.3",
-    "react-redux": "^7.2.0",
-    "react-scroll": "^1.8.1",
-    "react-select-search": "^2.0.4",
-    "redaxios": "^0.3.0",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0",
+    "@artsy/fresnel": "^6.1.0",
+    "@loadable/component": "^5.15.3",
+    "clsx": "^2.0.0",
+    "next": "^13.4.19",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-glider": "^4.0.2",
+    "react-lazyload": "^3.2.0",
+    "react-modal-video": "^2.0.1",
+    "react-redux": "^8.1.2",
+    "react-scroll": "^1.8.9",
+    "react-select-search": "^4.1.6",
+    "redaxios": "^0.5.1",
+    "redux": "^4.2.1",
+    "redux-thunk": "^2.4.2",
     "use-dark-mode": "^2.3.1"
   },
   "devDependencies": {
-    "@next/bundle-analyzer": "^9.5.3",
-    "@svgr/webpack": "^5.4.0",
+    "@next/bundle-analyzer": "^13.4.19",
+    "@svgr/webpack": "^8.1.0",
     "@zeit/next-source-maps": "0.0.3",
-    "cross-env": "^7.0.2",
-    "eslint": "^7.4.0",
-    "eslint-config-next": "^11.0.1",
+    "cross-env": "^7.0.3",
+    "eslint": "^8.48.0",
+    "eslint-config-next": "^13.4.19",
     "redux-devtools-extension": "^2.13.8",
     "redux-logger": "^3.0.6"
   }
